+++ packages/kubricate/src/index.ts
@@ -1,1 +1,1 @@
-export * from './lib/index.js';
+export * from './config.js';
