--- packages/kubricate/src/lib/color.ts
@@ -1,5 +1,0 @@
-import ansis from "ansis";
-
-export function colorCommand(string: string){
-  return ansis.blue(ansis.bold(string));
-}
