--- packages/kubricate/src/lib/index.ts
@@ -1,2 +1,0 @@
-export * from './config.js';
-export * from './color.js';
